Eliminated Funky, due to Swift 3 incompatibility
Eliminated usage of the Funky library, since it (still) hasn't been updated for Swift 3.

It's worth noting that the Swift 3 syntax for the things that Funky was doing here is surprisingly similar to Funky's syntax, and most cases more concise.  _It's plausible that the authors of Funky have abandoned work on their lib because Swift 3 added enough new functional-programming stuff to satisfy these types of workflows._

* Removed Funky from the `Podfile` & did a `pod update`.
* Removed `import Funky`s.
* Replaced `findWhere(«collection») { … }` usage with `«collection».index { … }`.
* Replaced `«collection» |> reducer(…) { … }` usage with `«collection».reduce(…) { … }`.
* Replaced `«collection» |> selectArray { … } |> toSet` usage with `Set(«collection».filter { … })`.